+++ stream-chat-android-ui-common/src/main/kotlin/io/getstream/chat/android/ui/common/feature/channel/attachments/ChannelAttachmentsViewEvent.kt
@@ -1,0 +1,31 @@
+/*
+ * Copyright (c) 2014-2025 Stream.io Inc. All rights reserved.
+ *
+ * Licensed under the Stream License;
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    https://github.com/GetStream/stream-chat-android/blob/main/LICENSE
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.getstream.chat.android.ui.common.feature.channel.attachments
+
+import io.getstream.result.Error
+
+/**
+ * The different types of events that can occur in the channel attachments view.
+ */
+public sealed interface ChannelAttachmentsViewEvent {
+    /**
+     * The error event that occurred while loading more attachments.
+     *
+     * @param error The error that occurred during the loading more process.
+     */
+    public data class LoadMoreError(val error: Error) : ChannelAttachmentsViewEvent
+}